Project structure is changed. Now every Resource subclass is seprated.
I have seprated some functions and classes from the main file. This is
jist a first step to project changes. The next step is to refactor out
the json 'rendering'. And to add different swagger versions.
