+++ flask_restful_swagger/__init__.py
@@ -1,9 +1,14 @@
 # -*- coding: utf-8 -*-
 
+import os
+
 __version__ = '1.0.0'
 
 registry = {
-  'models': {}
+    'models': {}
 }
 
 api_spec_static = ''
+
+root_path = os.path.dirname(__file__)
+templates = {}
