+++ flask_restful_swagger/resources/__init__.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+
+from __future__ import absolute_import
+
+from .swagger_registry import SwaggerRegistry
+from .resource_lister import ResourceLister
+from .static_files import StaticFiles
+from .swagger_resource import SwaggerResource, SwaggerResourceMeta
+
+__author__ = 'sobolevn'
+
